Bug 1841896 - Add bug reproduction for helper_fission_tap
During porting work on helper_fission_tap we found a bug in
the event retargeting. This page aims to reproduce it.
When clicking outside of the iframe but very close to the
button that is inside the iframe, the event should be
retargeted to the button. But it is not in every case.